+++ src/components/basic_table/in_memory_table.js
@@ -4,7 +4,6 @@ import {
   EuiBasicTable,
   ColumnType,
   SelectionType,
-  NoItemsMessageType
 } from './basic_table';
 import {
   defaults as paginationBarDefaults
@@ -23,7 +22,7 @@ const InMemoryTablePropTypes = {
   columns: PropTypes.arrayOf(ColumnType).isRequired,
   items: PropTypes.array,
   loading: PropTypes.bool,
-  message: NoItemsMessageType,
+  message: PropTypes.node,
   error: PropTypes.string,
   search: PropTypes.oneOfType([PropTypes.bool, PropTypes.shape({
     defaultQuery: QueryType,